requirements.txt
PyGithub
functions-framework